Coffee-Likod/requirements.txt
flask
flask-cors
pillow
ultralytics
numpy